requirements.txt
dash==1.20.0
dash-bootstrap-components==0.12.2
PyYAML==5.4.1
